voucher apply

+++ app/src/main/java/com/nt118/joliecafe/models/Voucher.kt
@@ -11,6 +11,6 @@ data class Voucher(
     val endDate: String,
     val type: String,
     val condition: Int,
-    val discountPercent: Int,
+    val discountPercent: Double,
     val quantity: Int
 )
